requirements.txt
numpy
pillow
scikit-image
scipy
httptools
uvloop
websockets
gunicorn
uvicorn
tqdm
fastapi
email-validator
aiofiles
matplotlib
